% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Growth.R
\name{ageCommunity}
\alias{ageCommunity}
\title{Updates the parameter file to a designated age}
\usage{
ageCommunity(base.params, age, tAge, growth, cover)
}
\arguments{
\item{base.params}{A parameter file}

\item{age}{The number of years since last fire}

\item{tAge}{The age of trees at the commencement of the run}

\item{growth}{A dataframe with the six fields:
Species - Name of the species consistent with other tables
max - Maximum plant height (m)
rate    - A constant describing the rate of growth for a Chapman Richards function}

\item{cover}{A dataframe with the fields:
Species - Name of the species consistent with other tables
constant - Mean plant separation (m) that does not change with age
exp_a    - The first constant in an exponential function describing plant separation with tsf
exp_b    - The second constant in an exponential function describing plant separation with tsf}
}
\value{
dataframe
}
\description{
Updates the parameter file to a designated age
}
