% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Growth.R
\name{applyGrowth}
\alias{applyGrowth}
\title{Updates a parameter file with the growth parameters modelled for a species}
\usage{
applyGrowth(Param, sp, current)
}
\arguments{
\item{Param}{A parameter file}

\item{sp}{The name of the species being modelled}

\item{current}{A dataframe with one row and the 5 fields
hp, ht, he, hc, w, clumpDiameter, clumpSeparation; values in m}
}
\value{
dataframe
}
\description{
Note: w is currently not changed due to an error:
"Error in .match_param(param, section, no.match.error = TRUE, single = TRUE) :
w matches more than one parameter"
}
