% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Fauna.R
\name{arboreal}
\alias{arboreal}
\title{Fire risk for an exposed arboreal animal}
\usage{
arboreal(
  Surf,
  Plant,
  percentile = 0.5,
  Height = 1,
  low = 1,
  high = 50,
  var = 10,
  Pressure = 1013.25,
  Altitude = 0,
  RH = 0.51,
  Class = "mammalia",
  Dimension = 0.1,
  Area = 0.2,
  protection = 0.02,
  count = 66,
  fibre = 0.01,
  Specific_heat = 2.5,
  skinCp = 3.5,
  skinK = 2,
  objectTemp = 38,
  Shape = "Flat",
  updateProgress = NULL
)
}
\arguments{
\item{Surf}{The dataframe 'runs' exported from Monte Carlos as 'Summary.csv'}

\item{Plant}{The dataframe 'IP' exported from Monte Carlos as 'IP.csv'.}

\item{percentile}{defines which heating statistics are used for each second, from 0 (min) to 1 (max)}

\item{Height}{The height directly ove ground (m) at which the species is expected to shelter from a fire.}

\item{low}{The closest horizontal distance between the flame origin and the point (m)}

\item{high}{The furthest horizontal distance between the flame origin and the point (m)}

\item{var}{The angle in degrees that the plume spreads above/below a central vector}

\item{Pressure}{Sea level atmospheric pressure (hPa)}

\item{Altitude}{Height above sea level (m)}

\item{RH}{The relative humidity (0-1)}

\item{Class}{Class of animal. Allowable values are "mammalia", "aves", "amphibia" and "reptilia"}

\item{Dimension}{The "Characteristic length" of the animal (m)}

\item{Area}{The surface area of the animal (m^2)}

\item{protection}{The thickness of fur, feather or scales covering the animal (m)}

\item{count}{The number of fibres per square mm}

\item{fibre}{The mean fibre diameter of hairs in mm}

\item{Specific_heat}{The specific heat of the fur, feather or scale material only (kJ/kg/deg C)}

\item{skinCp}{The specific heat of the animal skin (kJ/kg/C)}

\item{skinK}{Thermal conductivity of the animal skin (W/m/C)}

\item{objectTemp}{The body temperature of the animal (deg C)}

\item{Shape}{The approximate shape of the animal - either "Flat", "Sphere", or "Cylinder"}

\item{updateProgress}{Progress bar for use in the dashboard}
}
\value{
dataframe
}
\description{
Calculates the degree of injury or likelihood of mortality to an animal
caused by an approaching fire front
}
\details{
Utilises the output tables from 'threat' and 'radiation', and adds to these
the Reynolds Number, heat transfer coefficients, Newton's convective energy transfer coefficient,
and the temperature of the object each second.

Reynolds Number utilises a standard formulation (e.g. Gordon, N. T., McMahon, T. A. & Finlayson, B. L.
Stream hydrology: an introduction for ecologists. (Wiley, 1992))

Convective heat transfer coefficients use the widely adopted formulations of
Williams, F. A. Urban and wildland fire phenomenology. Prog. Energy Combust. Sci. 8, 317–354 (1982),
and Drysdale, D. An introduction to fire dynamics. (John Wiley and Sons, 1985)
utilising a Prandtl number of 0.7.

Mammal pelage is given a standardised emissivity of 0.86, based on:
McGowan, N. E., Scantlebury, D. M., Maule, A. G. & Marks, N. J.
Measuring the emissivity of mammal pelage. Quant. Infrared Thermogr. J. 6733, 1–9 (2018).

Skin Cp is set to 3.5, averaged from Duck, F. A.
Physical properties of tissues: a comprehensive reference book. (Elsevier Science, 1990).

Finds furDensity - the density of fur or feathers on the animal using standard values of 66 fibres per mm2,
fibre diameter of 0.01mm (10μm), and α-keratin density of 1300 kg.m-3. Fibre count averaged from
Liwanag, H. E. M., Berta, A., Costa, D. P., Abney, M. & Williams, T. M.
Morphological and thermal properties of mammalian insulation: the evolution of fur for aquatic living.
Biol. J. Linn. Soc. 106, 926–939 (2012)

Finds thermal dose from all heat inputs, using the formula from
Ciesielski, M., Mochnacki, B. & Szopa, R. Numerical modeling of biological tissue heating.
Admissible thermal dose. Sci. Res. Inst. Math. Comput. Sci. 10, 11–20 (2011).
}
