% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Flora.R
\name{bole}
\alias{bole}
\title{Air temperature above ambient at the tree bole behind the flame front}
\usage{
bole(lengthSurface = 2, residence = 300, depth = 0.05, h = 0.1,
  surfDecl = 10, t = 1)
}
\arguments{
\item{lengthSurface}{Surface flame length in m}

\item{residence}{Surface flame residence in seconds}

\item{depth}{Depth of surface litter in m}

\item{h}{Height of the point being measured}

\item{surfDecl}{Constant for the rate of decline in flame length}

\item{t}{Time in seconds behind the flame front}
}
\value{
temperature
}
\description{
Dynamic air temperature at bole height, declining flame length exponentially
}
\details{
Air temperature is modelled from dynamic flame segments using
Weber R.O., Gill A.M., Lyons P.R.A., Moore P.H.R., Bradstock R.A., Mercer G.N. (1995)
Modelling wildland fire temperatures. CALMScience Supplement, 4, 23–26.

pAlphas is set to bole height - depth of surface litter
}
