% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Flora.R
\name{cambium}
\alias{cambium}
\title{Finds radial bole necrosis depth}
\usage{
cambium(Surf, Plant, percentile = 0.95, Height = 0.1,
  woodDensity = 700, barkDensity = 500, bark = 0.04, comBark = 700,
  resBark = 45, cambThick = 0.01, xylemThick = 0.01, RH = 0.2,
  moisture = 1, bMoisture = 0.2, distance = 50, trail = 100,
  var = 10, diameter = 20, Pressure = 1013.25, Altitude = 0,
  startTemp = 25, necT = 60, surfDecl = 10, updateProgress = NULL)
}
\arguments{
\item{Surf}{The dataframe 'runs' exported from Monte Carlos as 'Summary.csv'}

\item{Plant}{The dataframe 'IP' exported from Monte Carlos as 'IP.csv'.}

\item{percentile}{defines which heating statistics are used for each second, from 0 (min) to 1 (max)}

\item{Height}{The height on the bole directly over ground (m)}

\item{woodDensity}{The density of wood in the tree or log housing the hollow (kg/m3)}

\item{barkDensity}{The density of bark in the tree or log housing the hollow (kg/m3)}

\item{bark}{The thickness of bark on the thinnest side of the hollow (m)}

\item{comBark}{Temperature directly under the burning bark (C)}

\item{resBark}{Flame residence in the tree bark (s)}

\item{cambThick}{Thickness of the tree cambium (m)}

\item{xylemThick}{Thickness of the active xylem (m)}

\item{RH}{The relative humidity (0-1)}

\item{moisture}{The proportion oven-dry weight of moisture in the wood}

\item{bMoisture}{The proportion oven-dry weight of moisture in the bark}

\item{distance}{The furthest horizontal distance between the flame origin and the point (m)}

\item{trail}{The number of seconds to continue modelling after all flames have extinguished}

\item{var}{The angle in degrees that the plume spreads above/below a central vector}

\item{diameter}{depth of the litter layer (mm)}

\item{Pressure}{Sea level atmospheric pressure (hPa)}

\item{Altitude}{Height above sea level (m)}

\item{startTemp}{The starting temperature of wood and bark (deg C)}

\item{necT}{Temperature of necrosis (deg C)}

\item{surfDecl}{adusts the rate at which surface flame length declines after the front}

\item{updateProgress}{Progress bar for use in the dashboard}
}
\value{
dataframe
}
\description{
Depth to which the cambium of a tree recieves lethal heating
}
\details{
Utilises the output tables from 'threat' and 'radiation', and adds to these
the Reynolds Number, heat transfer coefficients, Newton's convective energy transfer coefficient,
and the temperature of the object each second.

Reynolds Number utilises a standard formulation (e.g. Gordon, N. T., McMahon, T. A. & Finlayson, B. L.
Stream hydrology: an introduction for ecologists. (Wiley, 1992))

Convective heat transfer coefficients use the widely adopted formulations of
Williams, F. A. Urban and wildland fire phenomenology. Prog. Energy Combust. Sci. 8, 317–354 (1982),
and Drysdale, D. An introduction to fire dynamics. (John Wiley and Sons, 1985)
utilising a Prandtl number of 0.7.

Heat is transferred into the bark and timber using Fourier's Law

Thermal conductivity of bark is modelled as per Martin, R. E.
Thermal properties of bark. For. Prod. J. 13, 419–426 (1963)

Specific heat of bark is modelled using Kain, G., Barbu, M. C., Hinterreiter, S., Richter, K. & Petutschnigg, A.
Using bark as a heat insulation material. BioResources 8, 3718–3731 (2013)

Thermal conductivity of wood is modelled using an approach from Kollmann, F. F. P. & Cote, W. A.
Principles of wood science and technology I. Solid wood. (Springer-Verlag, 1968)

Evaporates water at 100 degrees C

Specific heat of wood is derived from an established empirical relationship in Volbehr, B.
Swelling of wood fiber. PhD Thesis. (University of Kiel, 1896)

Continues heating of the bole in the wake of the front for the duration of the surface fire for a period determined using
Burrows, N. D. Flame residence times and rates of weight loss of eucalypt forest fuel particles.
Int. J. Wildl. Fire 10, 137–143 (2001). Flame lengths are decreased exponentially over this period

Bark is assumed to ignite, and burn for an average of resBark seconds, with the default value of 45s used as a mean for
figure 6c in Penman, T. D., Cawson, J. G., Murphy, S. & Duff, T. J.
Messmate Stringybark: bark ignitability and burning sustainability in relation to fragment dimensions,
hazard and time since fire. Int. J. Wildl. Fire 26, 866–876 (2017).

Heats an area of 0.01m2
}
