% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Earth.R
\name{cpSoil}
\alias{cpSoil}
\title{Specific heat of soil}
\usage{
cpSoil(temp = 300, texture = "loam", peat = 0.2, moisture = 0.3)
}
\arguments{
\item{temp}{soil temperature in K}

\item{texture}{Soil texture. Allowable values are: "sand", "loamy sand", "sandy loam", "sandy clay loam",
"sand clay", "loam", "clay loam", "silt loam", "clay", "silty clay", "silty clay loam", "silt"}

\item{peat}{Organic proportion of the soil}

\item{moisture}{Water proportion ODW of the soil}
}
\value{
Specific heat of soil (J/kg.K)
}
\description{
Finds volumetric specific heat from the mineral, organic and water components of the soil
}
\details{
Specific heats of soil components estimated from figures 108 & 111 in
Farouki, O. Thermal properties of soils. (Trans Tech, 1981).

Water specific heat 4185 J/kg.K
}
