% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Earth.R
\name{denSoil}
\alias{denSoil}
\title{Soil density}
\usage{
denSoil(texture = "loam")
}
\arguments{
\item{texture}{Soil texture. Allowable values are: "sand", "loamy sand", "sandy loam", "sandy clay loam",
"sand clay", "loam", "clay loam", "silt loam", "clay", "silty clay", "silty clay loam", "silt"}
}
\description{
Finds soil density from texture
}
\details{
Porosity of soils taken from Rawls, W. J., Brakensiek, D. L. & Saxton, K. E.
Estimation of soil water properties. Transactions of the ASAE 25, 1316–1320 & 1328 (1982).

Density equation is from Peters-Lidard, C. D., Blackburn, E., Liang, X. & Wood, E. F.
The effect of soil thermal conductivity parameterization on surface energy fluxes and temperatures.
J. Atmos. Sci. 55, 1209–1224 (1998).
}
