% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Dashboard.R
\name{drivers}
\alias{drivers}
\title{Models fire behaviour across ranged variables}
\usage{
drivers(base.params, db.path = "out_mc.db", jitters, windMin, windReps,
  windStep, moistureMultiplier, moistureSD, moistureRange, heightSD,
  heightRange, leafVar, updateProgress = NULL)
}
\arguments{
\item{base.params}{Input parameter file}

\item{db.path}{Name of the exported database}

\item{jitters}{Number of repetitions for each row in the weather table}

\item{windMin}{Lowest wind velocity tested (km/h)}

\item{windReps}{Number of wind speeds tested}

\item{windStep}{Gap (km/h) between wind steps}

\item{moistureMultiplier}{Multiplies all LFMC values by this number}

\item{moistureSD}{Standard deviation of moisture}

\item{moistureRange}{Truncates variability by +/- mean * range}

\item{heightSD}{Standard deviation of plant height}

\item{heightRange}{Truncates variability by +/- mean * range}

\item{leafVar}{Variation around input leaf dimensions, equivalent to l}

\item{updateProgress}{Progress bar for use in the dashboard}
}
\description{
Models fire behaviour across ranged variables
}
