% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Dashboard.R
\name{driversS}
\alias{driversS}
\title{Models fire behaviour across ranged variables}
\usage{
driversS(base.params, db.path = "out_mc.db", jitters, windMin, windReps,
  windStep, moistureMultiplier, moistureSD, moistureRange, Variation,
  leafVar, updateProgress = NULL)
}
\arguments{
\item{base.params}{Input parameter file}

\item{db.path}{Name of the exported database}

\item{jitters}{Number of repetitions for each row in the weather table}

\item{windMin}{Lowest wind velocity tested (km/h)}

\item{windReps}{Number of wind speeds tested}

\item{windStep}{Gap (km/h) between wind steps}

\item{moistureMultiplier}{Multiplies all LFMC values by this number}

\item{moistureSD}{Standard deviation of moisture}

\item{moistureRange}{Truncates variability by +/- mean * range}

\item{Variation}{A database of plant variability in traits, with the fields:
record - a unique, consecutively numbered identifier per site
species - the name of the species, which will call trait data from 'default.species.params'
stratum - numeric value from 1 to 4, counting from lowest stratum
Hs - Standard deviation of plant height variations
Hr - Truncates plant height variability by +/- Hr * height}

\item{leafVar}{Variation around input leaf dimensions, equivalent to l}

\item{updateProgress}{Progress bar for use in the dashboard}
}
\description{
Models fire behaviour across ranged variables
}
