% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Flora.R
\name{flora}
\alias{flora}
\title{Scorch height using an isotherm}
\usage{
flora(Surf, Plant, Param = Param, Test = 70)
}
\arguments{
\item{Surf}{The dataframe 'runs' exported from Monte Carlos as 'Summary.csv'}

\item{Plant}{The dataframe 'IP' exported from Monte Carlos as 'IP.csv'.}

\item{Param}{A parameter dataframe used for FRaME,
such as produced using readLegacyParamFile}

\item{Test}{The temperature of the flora, default 60 degC}
}
\value{
dataframe
}
\description{
Calculates the height to which vegetation will be consumed,
and the height to a designated temperature isotherm reached for one second.
}
\details{
Output fields are:
ht - scorch height from the surface flame (m)
hp - scorch height from burning plant (m)
Height - overall scorch height (m)
ns, e, m, c - height of consumption in each stratum (m)
b1, b2, b3, b4 - percentage of strata 1 (ns) to strata 4 (c) consumed
sc1, sc2, sc3, sc4 - percentage of strata 1 (ns) to strata 4 (c) scorched
d1, d2, d3, d4 - death (1) or survival (0) of standing foliage per stratum (50% or more scorch)
}
