% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Growth.R
\name{growPlants}
\alias{growPlants}
\title{Models plant height from time since fire}
\usage{
growPlants(Param, a, sp, stn, growth, age)
}
\arguments{
\item{Param}{A parameter file}

\item{a}{The number of the record to be modelled}

\item{sp}{The name of the species being modelled}

\item{stn}{The number of the stratum}

\item{growth}{A dataframe with the six fields:
Species - Name of the species consistent with other tables
max - Maximum plant height (m)
rate    - A constant describing the rate of growth for a Chapman Richards function}

\item{age}{The number of years since last fire}
}
\value{
dataframe
}
\description{
Uses either standard Chapman-Richards negative exponential or linear functions to grow hp
All other parameters are altered to maintain original proportions to hp
}
