% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Earth.R
\name{kSoil}
\alias{kSoil}
\title{Thermal conductivity of soil}
\usage{
kSoil(texture = "loam", saturation = 0.3, grain = "fine",
  unfrozen = 1)
}
\arguments{
\item{texture}{Soil texture. Allowable values are: "sand", "loamy sand", "sandy loam", "sandy clay loam",
"sand clay", "loam", "clay loam", "silt loam", "clay", "silty clay", "silty clay loam", "silt"}

\item{saturation}{Water saturation of the soil, value between 0.1 and 1}

\item{grain}{Allowable values are "fine" or "coarse"}

\item{unfrozen}{Proportion of soil unfrozen, between 0 and 1}
}
\value{
Thermal conductivity of soil (W/mK)
}
\description{
Model drawn from Johansen, O.
Thermal conductivity of soils. PhD Thesis (University of Trondheim, 1971)
Modified by Farouki, O. Thermal properties of soils. (Trans Tech, 1986)
}
\details{
Porosity taken from Rawls, W. J., Brakensiek, D. L. & Saxton, K. E.
Estimation of soil water properties. Transactions of the ASAE 25, 1316–1320 & 1328 (1982).
}
