% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Flora.R
\name{kWood}
\alias{kWood}
\title{Thermal conductivity of dry wood}
\usage{
kWood(T = 100, rhoW = 700, kAir = 0.026)
}
\arguments{
\item{T}{Air temperature (deg C)}

\item{rhoW}{Timber density (kg/m3)}

\item{kAir}{Thermal conductivity of the air at temperature T (W/M.K)}
}
\value{
Thermal conductivity of wood (W/mK)
}
\description{
Model drawn from Kollmann, F. F. P. & Cote, W. A.
Principles of wood science and technology I. Solid wood. (Springer-Verlag, 1968)
}
