% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Shelter.R
\name{LAI}
\alias{LAI}
\title{Finds the LAI for a horizontal slice of a community}
\usage{
LAI(base.params, yu = 100, yl = 0)
}
\arguments{
\item{base.params}{A parameter file}

\item{yu}{The upper height of the horizontal slice (m)}

\item{yl}{The lower height of the horizontal slice (m)}
}
\value{
Numeric value
}
\description{
Finds the LAI for a horizontal slice of a community
}
