% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Shelter.R
\name{LAIp}
\alias{LAIp}
\title{Finds the LAI for a horizontal slice of a plant}
\usage{
LAIp(base.params, sp = 1, yu = 100, yl = 0)
}
\arguments{
\item{base.params}{A parameter file}

\item{yu}{The upper height of the horizontal slice (m)}

\item{yl}{The lower height of the horizontal slice (m)}

\item{Sp}{The number of the species in the parameter file}
}
\value{
Numeric value
}
\description{
Finds the LAI for a horizontal slice of a plant
}
