% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/match_test_ctrl.R
\name{match_test_ctrl}
\alias{match_test_ctrl}
\title{Match ctrl units to identified test units}
\usage{
match_test_ctrl(cluster_formula, data, type = "group", method = "diff",
  test_name = "test", ctrl_name = "ctrl", replace = FALSE, m = NA,
  group_tol = 0.01)
}
\arguments{
\item{cluster_formula}{a formula of the type trt ~ covariates}

\item{data}{a data.frame where each row is a unit and the columns are observations of the covariates to match on.  Must at least contain a trt variable.}

\item{type}{the type of analysis to perform, diff or ratio}

\item{method}{the method of control matches, group, 1-1, or 1-m}

\item{test_name}{the name of the test level in the trt variable}

\item{ctrl_name}{the name of the control level in the trt variable}

\item{replace}{should the control groups be replaced in the pool after they are matched to a test unit}

\item{m}{the number of ctrl units should be matched to each test unit}

\item{group_tol}{the relative tolerance of the group matching algorithm.  One stopping criteria is that the distance between the full control group and the test group is less than group_tol times the original distance.}
}
\value{
an object of matchedTestCtrl
}
\description{
Inspired by the matchControl function in e1071
}
\examples{
set.seed(123)
df <- data.frame(id = factor(LETTERS),
                 trt = factor(c(rep("test", 4), rep("ctrl", 22))),
                 metric1_time1 = rnorm(26),
                 metric1_time2 = rnorm(26),
                 metric1_time3 = rnorm(26),
                 metric2_time1 = rnorm(26),
                 metric2_time2 = rnorm(26),
                 metric2_time3 = rnorm(26))
m1 <- match_test_ctrl(paste("trt ~", paste(
  paste("metric", rep(1:2, each = 3),
        "_time", rep(1:3, times = 2),
        sep = ""), collapse = " + ")),
  data = df,
  type = "1-1", method = "diff",
  test_name = "test", ctrl_name = "ctrl",
  replace = FALSE,
  m = NA, group_tol = 0.01)
}
