% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Growth.R
\name{olson}
\alias{olson}
\title{Models the weight of the o_horizon from time since fire and
updates param file}
\usage{
olson(base.params, growth, age)
}
\arguments{
\item{base.params}{Parameter input table}

\item{growth}{A dataframe with the six fields:
species - Name of the species consistent with other tables
max - Maximum plant height (m)
rate    - A constant describing the rate of growth for a Chapman Richards function}

\item{age}{The number of years since last fire}
}
\value{
dataframe
}
\description{
olson negative exponential function
}
