% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Radiation.R
\name{phi}
\alias{phi}
\title{Configuration or view factor}
\usage{
phi(Fl, Fa, S, D, H)
}
\arguments{
\item{Fl}{Flame length (metres)}

\item{Fa}{Flame angle (radians)}

\item{S}{Slope (degrees)}

\item{D}{Distance to the receiver (metres)}

\item{H}{Height of the receiver (metres)}
}
\value{
Phi - the configuration factor
}
\description{
Calculates the configuration factor for radiative heat transfer as used
in AS3959 for bushfire risk assessment of built structures
}
