% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ensembles.R
\name{plantVar}
\alias{plantVar}
\title{Randomly modifies plant traits within defined ranges for non-deterministic predictions}
\usage{
plantVar(base.params, Strata, Species, l = 0.1, Ms = 0.01, Pm = 1,
  Mr = 1.001, Hs = 0.2, Hr = 1.41)
}
\value{
dataframe
}
\description{
Randomly modifies plant traits within defined ranges for non-deterministic predictions
}
