% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ensembles.R
\name{plantVarS}
\alias{plantVarS}
\title{Randomly modifies plant traits within defined ranges for non-deterministic predictions
Differs from plantVar by modifying individual species by their own rules}
\usage{
plantVarS(base.params, Strata, Species, Variation, a, l = 0.1,
  Ms = 0.01, Pm = 1, Mr = 1.001)
}
\arguments{
\item{base.params}{Parameter input table}

\item{Strata}{Strata descriptor table output by the function 'strata'}

\item{Species}{Species descriptor table output by the function 'species'}

\item{Variation}{A database of plant variability in traits, with the fields:
record - a unique, consecutively numbered identifier per site
species - the name of the species, which will call trait data from 'default.species.params'
stratum - numeric value from 1 to 4, counting from lowest stratum
Hs - Standard deviation of plant height variations
Hr - Truncates plant height variability by +/- Hr * height}

\item{l}{Variation around input leaf dimensions}

\item{Ms}{Standard deviation of LFMC}

\item{Pm}{Multiplier of mean LFMC}

\item{Mr}{Truncates LFMC variability by +/- Pm * LFMC}
}
\value{
dataframe
}
\description{
Randomly modifies plant traits within defined ranges for non-deterministic predictions
Differs from plantVar by modifying individual species by their own rules
}
