% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_impactResult.R
\name{plot.impactResult}
\alias{plot.impactResult}
\title{Plot impact Result Histogram}
\usage{
\method{plot}{impactResult}(x, ...)
}
\arguments{
\item{x}{impactResult object}

\item{...}{parameters passed to stats::hist}
}
\description{
Plot impact Result Histogram
}
\examples{
a <- 4
}
