% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_matchedTestCtrl.R
\name{plot.matchedTestCtrl}
\alias{plot.matchedTestCtrl}
\title{Plot a matchedTestCtrl object using Principal components}
\usage{
\method{plot}{matchedTestCtrl}(x, ...)
}
\arguments{
\item{x}{a matchedTestCtrl object}

\item{...}{other parameters passed to internal methods (future use)}
}
\value{
a ggplot object
}
\description{
Plot a matchedTestCtrl object using Principal components
}
\examples{
df <- data.frame(id = factor(LETTERS),
                 trt = factor(c(rep("test", 4), rep("ctrl", 22))),
                 metric1_time1 = c(0,5,10,15,1,6,11,16:34),
                 metric1_time2 = c(0,5,10,15,1,6,11,16:34),
                 metric1_time3 = c(0,5,10,15,1,6,11,16:34),
                 metric2_time1 = rnorm(26),
                 metric2_time2 = rnorm(26),
                 metric2_time3 = rnorm(26))

m3 <- match_test_ctrl(paste("trt ~", paste(
  paste("metric", rep(1:2, each = 3),
        "_time", rep(1:3, times = 2),
        sep = ""), collapse = " + ")),
  data = df,
  type = "1-1", method = "diff",
  test_name = "test", ctrl_name = "ctrl",
  replace = TRUE,
  m = NA, group_tol = 0.01)

plot(m3)
}
