% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_impact_data_step_change.R
\name{plot_impact_data_step_change}
\alias{plot_impact_data_step_change}
\title{Plot the impact data step change}
\usage{
plot_impact_data_step_change(dat, method = "diff", type = "raw")
}
\arguments{
\item{dat}{impact data}

\item{method}{diff or ratio}

\item{type}{raw or corrected}
}
\value{
ggplot2 object
}
\description{
Plot the impact data step change
}
\examples{
a <- 4
}
