% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Dashboard.R
\name{probFire}
\alias{probFire}
\title{Models probabilistic fire behaviour}
\usage{
probFire(base.params, db.path = "out_mc.db", jitters, slope, slopeSD,
  slopeRange, temp, tempSD, tempRange, DFMC, DFMCSD, DFMCRange, wind,
  windSD, windRange, moistureMultiplier, moistureSD, moistureRange,
  heightSD, heightRange, leafVar, updateProgress = NULL)
}
\arguments{
\item{base.params}{Input parameter file}

\item{db.path}{Name of the exported database}

\item{jitters}{Number of repetitions for each row in the weather table}

\item{slope}{Mean slope (deg)}

\item{slopeSD}{Standard deviation of slope}

\item{slopeRange}{Truncates variability by +/- mean * range}

\item{temp}{Mean ambient temperature (deg.C)}

\item{tempSD}{Standard deviation of temp}

\item{tempRange}{Truncates variability by +/- mean * range}

\item{DFMC}{Mean DFMC (%ODW)}

\item{DFMCSD}{Standard deviation of DFMC}

\item{DFMCRange}{Truncates variability by +/- mean * range}

\item{wind}{Mean wind velocity (km/h)}

\item{windSD}{Standard deviation of wind velocity}

\item{windRange}{Truncates variability by +/- mean * range}

\item{moistureMultiplier}{Multiplies all LFMC values by this number}

\item{moistureSD}{Standard deviation of moisture}

\item{moistureRange}{Truncates variability by +/- mean * range}

\item{heightSD}{Standard deviation of plant height}

\item{heightRange}{Truncates variability by +/- mean * range}

\item{leafVar}{Variation around input leaf dimensions, equivalent to l}

\item{updateProgress}{Progress bar for use in the dashboard}
}
\value{
dataframe
}
\description{
Models probabilistic fire behaviour
}
