% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Shelter.R
\name{profileDet}
\alias{profileDet}
\title{Finds the deterministic vertical wind profile for a community}
\usage{
profileDet(base.params, slices = 10)
}
\arguments{
\item{base.params}{A parameter file}

\item{slices}{The number of vertical slices to measure}
}
\value{
table
}
\description{
Finds the deterministic vertical wind profile for a community
}
