% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Summaries.R
\name{repFlame}
\alias{repFlame}
\title{Representative flame descriptors}
\usage{
repFlame(IP)
}
\arguments{
\item{IP}{The dataframe res$IP}
}
\value{
dataframe
}
\description{
Summarises FRaME generated flame segments into a combined,
representative plant flame for each repId where plants ignited
}
