% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Shelter.R
\name{shelter}
\alias{shelter}
\title{Finds the deterministic vertical wind profile for a community, above a defined screen level}
\usage{
shelter(base.params, slices = 10, SL = 1.2)
}
\arguments{
\item{base.params}{A parameter file}

\item{slices}{The number of vertical slices to measure above the measurement point}

\item{height}{The height (m) of the lower wind measurement point}
}
\value{
table
}
\description{
Finds the deterministic vertical wind profile for a community, above a defined screen level
}
