% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Building.R
\name{siteBuilder}
\alias{siteBuilder}
\title{Builds the dataframe site.meta from input tables}
\usage{
siteBuilder(site, Structure, a)
}
\arguments{
\item{site}{A dataframe with the six fields:
record - a unique, consecutively numbered identifier per site
site - the name of each record
slope - slope in degrees
wind - velocity in km/h
temp - ambient temperature deg. C
dfmc - moisture content of fine dead fuels in whole numbers (eg 0.1 for 10%)
oHorizon - weight in t/ha of fine dead organic material forming the O horizon
fline - the fireline length in m}

\item{Structure}{A dataframe with the fields:
record - a unique, consecutively numbered identifier per site
site - a unique identifier per site
NS, El, Mid & Can - the mean separation between plants (m) per stratum
ns_e, ns_m, e_m, e_c, m_c - Logical field indicating whether plants in the stratum
on the left grow directly beneath those in the stratum on the right. Acceptable values
are t, f, or blank, where the outcome will be decided by the relative stratum heights.}

\item{a}{The site number for which to build the table}
}
\description{
Builds the dataframe site.meta from input tables
}
