% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ensembles.R
\name{spComb}
\alias{spComb}
\title{Models fire behaviour in each species of a parameter table,
summarising the combustibility from the length of flame divided
by the length of segment ignited}
\usage{
spComb(base.params)
}
\arguments{
\item{base.params}{A parameter file}
}
\value{
dataframe
}
\description{
Models fire behaviour in each species of a parameter table,
summarising the combustibility from the length of flame divided
by the length of segment ignited
}
