% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Summaries.R
\name{species}
\alias{species}
\title{Species descriptors from a param file}
\usage{
species(Param)
}
\arguments{
\item{Param}{A parameter dataframe used for FRaME,
such as produced using readLegacyParamFile}
}
\value{
dataframe of species descriptors
}
\description{
Finds dimensions and moisture of each species
}
