% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Building.R
\name{speciesBuilder}
\alias{speciesBuilder}
\title{Builds the dataframe species.values from input tables}
\usage{
speciesBuilder(Flora, site, a)
}
\arguments{
\item{Flora}{A dataframe with the fields:
record - a unique, consecutively numbered identifier per site
species - the name of the species, which will call trait data from 'default.species.params'
moisture - the moisture content of the species in whole numbers (eg 1 for 100% ODW)
stratum - numeric value from 1 to 4, counting from lowest stratum
comp - % composition or count of that species in the stratum. If absent, all species will be considered equally
base, he, ht, top & w - canopy dimensions for that species (m). he and ht are optional
clump - mean ratio of clump diameter to crown diameter
openness - ratio of gap to clump size}

\item{site}{A dataframe with the six fields:
record - a unique, consecutively numbered identifier per site
site - the name of each record
slope - slope in degrees
wind - velocity in km/h
temp - ambient temperature deg. C
dfmc - moisture content of fine dead fuels in whole numbers (eg 0.1 for 10%)
oHorizon - weight in t/ha of fine dead organic material forming the O horizon
fline - the fireline length in m}

\item{a}{The site number for which to build the table}
}
\description{
Builds the dataframe species.values from input tables
}
