% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ensembles.R
\name{specPoint}
\alias{specPoint}
\title{Selects random species from the available list, weighted by their frequency
Modifies a parameter table to the shortened list}
\usage{
specPoint(base.params, Structure, a)
}
\arguments{
\item{base.params}{A parameter file}

\item{Structure}{A dataframe with the fields:
record - a unique, consecutively numbered identifier per site
site - a unique identifier per site
NS, El, Mid & Can - the mean separation between plants (m) per stratum
ns_e, ns_m, e_m, e_c, m_c - Logical field indicating whether plants in the stratum
on the left grow directly beneath those in the stratum on the right. Acceptable values
are t, f, or blank, where the outcome will be decided by the relative stratum heights.
nsR, eR, mR, cR - maximum species richness recorded for each stratum}
}
\value{
dataframe
}
\description{
Selects random species from the available list, weighted by their frequency
Modifies a parameter table to the shortened list
}
