% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Summaries.R
\name{strata}
\alias{strata}
\title{Stratum descriptors from a param file}
\usage{
strata(Param)
}
\arguments{
\item{Param}{A parameter dataframe used for FRaME,
such as produced using readLegacyParamFile}
}
\value{
dataframe of stratum descriptors
}
\description{
For each stratum, finds mean crown width, plant separation, and number of species
}
