% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Building.R
\name{strataBuilder}
\alias{strataBuilder}
\title{Builds the dataframe strata.meta from input tables}
\usage{
strataBuilder(Structure, Flora, a)
}
\arguments{
\item{Structure}{A dataframe with the fields:
record - a unique, consecutively numbered identifier per site
site - the name of each record
NS, El, Mid & Can - the mean separation between plants (m) per stratum
ns_e, ns_m, e_m, e_c, m_c - Logical field indicating whether plants in the stratum
on the left grow directly beneath those in the stratum on the right. Acceptable values
are t, f, or blank, where the outcome will be decided by the relative stratum heights.}

\item{Flora}{A dataframe with the fields:
record - a unique, consecutively numbered identifier per site
species - the name of the species, which will call trait data from 'default.species.params'
moisture - the moisture content of the species in whole numbers (eg 1 for 100% ODW)
stratum - numeric value from 1 to 4, counting from lowest stratum
comp - % composition or count of that species in the stratum. If absent, all species will be considered equally
base, he, ht, top & w - canopy dimensions for that species (m). he and ht are optional
clump - mean ratio of clump diameter to crown diameter
openness - ratio of gap to clump size}

\item{a}{The record number for which to build the table}
}
\description{
Builds the dataframe strata.meta from input tables
}
