% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Summaries.R
\name{stratum}
\alias{stratum}
\title{Summary table of stratum results}
\usage{
stratum(flames, sites, ros, surf)
}
\arguments{
\item{flames}{The dataframe res$FlameSummaries}

\item{sites}{The dataframe res$Sites}

\item{ros}{The dataframe res$ROS}

\item{surf}{The dataframe produced by the extension surf}
}
\value{
stratum dataframe
}
\description{
Summarises FRaME generated fire behaviour by stratum and RepId
}
