% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Summaries.R
\name{summary}
\alias{summary}
\title{Summary table of fire behaviour}
\usage{
summary(stratum, surf)
}
\arguments{
\item{stratum}{The dataframe stratum}

\item{surf}{The dataframe surf}
}
\value{
summary dataframe
}
\description{
Summarises FRaME generated fire behaviour by RepId
}
