% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Summaries.R
\name{surf}
\alias{surf}
\title{Summary table of surface results}
\usage{
surf(surface)
}
\arguments{
\item{surface}{The dataframe res$SurfaceResults}
}
\value{
dataframe
}
\description{
Summarises FRaME generated surface fire behaviour by RepId
}
