% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Growth.R
\name{susp}
\alias{susp}
\title{Models the packing of materials in a suspended layer}
\usage{
susp(
  base.params,
  a,
  suspNS = "suspNS",
  Flora,
  growth,
  default.species.params,
  age,
  density = 300
)
}
\arguments{
\item{base.params}{Parameter input table}

\item{a}{The number of the record to be modelled}

\item{suspNS}{Name of the fuel in the suspended layer}

\item{Flora}{A dataframe with the fields:
record - a unique, consecutively numbered identifier per site
species - the name of the species, which will call trait data from 'default.species.params'
moisture - the moisture content of the species in whole numbers (eg 1 for 100% ODW)
stratum - numeric value from 1 to 4, counting from lowest stratum
comp - % composition or count of that species in the stratum. If absent, all species will be considered equally
base, he, ht, top & w - canopy dimensions for that species (m). he and ht are optional
clump - mean ratio of clump diameter to crown diameter
openness - ratio of gap to clump size}

\item{growth}{A dataframe with the six fields:
species - Name of the species consistent with other tables
max - Maximum plant height (m)
rate    - A constant describing the rate of growth for a Chapman Richards function
aQ - First constant in a quadratic equation
bQ - Second constant in a quadratic equation
cQ - Third constant in a quadratic equation
aLin - First constant in a linear equation
bLin - Second constant in a linear equation
Function prioritises growth equations from left to right}

\item{default.species.params}{Leaf traits database}

\item{age}{The number of years since last fire}

\item{density}{Wood density (kg/m3)}
}
\value{
dataframe
}
\description{
Based on a negative exponential function;updates base.params
}
