% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_level_shift.R
\name{test_level_shift}
\alias{test_level_shift}
\title{Test Level Shift in Test-Control Experiment}
\usage{
test_level_shift(dat, type = "group", method = "diff",
  val_name = "val", test_name = "test", ctrl_name = "ctrl",
  pre_name = "pre", post_name = "post", R = 1000)
}
\arguments{
\item{dat}{a dataframe with columns including pre_post, time, id, trt, val}

\item{type}{\code{group} for testing two groups of size m and n, \code{1-1} for testing matched pairs, \code{1-m} for testing matches between one test and m ctrl}

\item{method}{\code{diff} for normalizing by the difference in test=control and \code{ratio} for normalizing by the ratio of test / control}

\item{val_name}{the name the value column in the data.frame dat}

\item{test_name}{the name of the test group in the trt column}

\item{ctrl_name}{the name of the control group in the trt column}

\item{pre_name}{the name of the pre time period in the pre_post column}

\item{post_name}{the name of the post time period in the pre_post column}

\item{R}{the number of boostrap replicates}
}
\value{
impactResult object
}
\description{
Test Level Shift in Test-Control Experiment
}
\examples{
dat <- data.frame(trt = rep(c("test","ctrl"), each = 6),
                  pre_post = rep(c("pre","post"), each = 3, times = 2),
                  id = as.character(c(1,2,3,1,2,3,4,5,6,4,5,6)),
                  time = as.character(c(1,1,1,2,2,2,1,1,1,2,2,2)),
                  val = c(10,11,12,13,14,15,10,11,10,12,10,10.5))
test_result <- test_level_shift(dat, type = "group", method = "diff", R = 100)
}
