% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_response_change.R
\name{test_response_change}
\alias{test_response_change}
\title{Test a change in response between test and control}
\usage{
test_response_change(dat, type = "group", method = "diff",
  val_name = "val", test_name = "test", ctrl_name = "ctrl",
  pre_name = "pre", post_name = "post", test_formula = ifelse(type ==
  "group", .defaultFormula, .defaultFormula1_1), test_time = NA,
  R = 1000)
}
\arguments{
\item{dat}{a dataframe with columns including pre_post, time, id, trt, val}

\item{type}{\code{group} for testing two groups of size m and n, \code{1-1} for testing matched pairs, \code{1-m} for testing matches between one test and m ctrl}

\item{method}{\code{diff} for normalizing by the difference in test=control and \code{ratio} for normalizing by the ratio of test / control}

\item{val_name}{the name the value column in the data.frame dat}

\item{test_name}{the name of the test group in the trt column}

\item{ctrl_name}{the name of the control group in the trt column}

\item{pre_name}{the name of the pre time period in the pre_post column}

\item{post_name}{the name of the post time period in the pre_post column}

\item{test_formula}{the formula to be used to conduct the test}

\item{test_time}{the time at which to conduct the test}

\item{R}{the number of boostrap replicates}
}
\value{
impactResult object
}
\description{
Test a change in response between test and control
}
\examples{
set.seed(1976)
 dat <- data.frame(trt = rep(c("test","ctrl"), each = 12),
                   pre_post = rep(c("pre","post"), each = 6, times = 2),
                   id = c(rep(c("1","2","3"), times = 4), rep(c("4","5","6"), times = 4)),
                   time = rep(1:4, each = 3, times = 2),
                   val = rnorm(24, mean = 10, sd = 5))
test_result <- test_response_change(dat, R = 100)
}
