% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Building.R
\name{unitBuilder}
\alias{unitBuilder}
\title{Builds the dataframe species.units from input tables}
\usage{
unitBuilder(Flora, a)
}
\arguments{
\item{Flora}{A dataframe with the fields:
record - a unique, consecutively numbered identifier per site
species - the name of the species, which will call trait data from 'default.species.params'
moisture - the moisture content of the species in whole numbers (eg 1 for 100% ODW)
stratum - numeric value from 1 to 4, counting from lowest stratum
comp - % composition or count of that species in the stratum. If absent, all species will be considered equally
base, he, ht, top & w - canopy dimensions for that species (m). he and ht are optional
clump - mean ratio of clump diameter to crown diameter
openness - ratio of gaps between clumps to clump size}

\item{a}{The record number for which to build the table}
}
\description{
Builds the dataframe species.units from input tables
}
