% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_impact_data.R
\name{validate_impact_data}
\alias{validate_impact_data}
\title{Validate the input data for the impact package}
\usage{
validate_impact_data(dat)
}
\arguments{
\item{dat}{an input data.frame}
}
\value{
TRUE if the data.frame is valid or an exception otherwise
}
\description{
Validate the input data for the impact package
}
\examples{
dat <- data.frame(trt = rep(c("test","ctrl"), each = 6),
                  pre_post = rep(c("pre","post"), each = 3, times = 2),
                  id = as.character(c(1,2,3,1,2,3,4,5,6,4,5,6)),
                  time = as.character(c(1,1,1,2,2,2,1,1,1,2,2,2)),
                  val = c(10,11,12,13,14,15,10,11,10,12,10,10.5))
validate_impact_data(dat)
}
