% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ensembles.R
\name{weatherSet}
\alias{weatherSet}
\title{Updates parameter files with weather from a dataset,
then models fire from non-deterministic plant parameters}
\usage{
weatherSet(base.params, weather, db.path = "out_mc.db", jitters = 10,
  l = 0.1, Ms = 0.01, Pm = 1, Mr = 1.001, Hs = 0.2, Hr = 1.41,
  updateProgress = NULL)
}
\arguments{
\item{base.params}{Parameter input table}

\item{weather}{A dataframe with the four fields:
tm - Sequential numbering of the records
T - Air temperature (deg C)
W - Wind velocity (km/h)
DFMC - Dead fuel moisture content (proportion ODW)}

\item{db.path}{Name of the exported database}

\item{jitters}{Number of repetitions}

\item{l}{Variation around input leaf dimensions}

\item{Ms}{Standard deviation of LFMC}

\item{Pm}{Multiplier of mean LFMC}

\item{Mr}{Truncates LFMC variability by +/- Pm * LFMC}

\item{Hs}{Standard deviation of plant height variations}

\item{Hr}{Truncates plant height variability by +/- Hr * height}

\item{updateProgress}{Progress bar for use in the dashboard}
}
\value{
dataframe
}
\description{
Updates parameter files with weather from a dataset,
then models fire from non-deterministic plant parameters
}
