% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Shelter.R
\name{windProfileP}
\alias{windProfileP}
\title{Calculates a non-deterministic wind profile on parallel cores
Useful when running a large number of reps}
\usage{
windProfileP(base.params, Variation, reps = 1000, slices = 100,
  l = 0.1, Ms = 0.01, Pm = 1, Mr = 1.001)
}
\arguments{
\item{base.params}{Parameter input table}

\item{Variation}{A database of plant variability in traits, with the fields:
record - a unique, consecutively numbered identifier per site
species - the name of the species, which will call trait data from 'default.species.params'
stratum - numeric value from 1 to 4, counting from lowest stratum
Hs - Standard deviation of plant height variations
Hr - Truncates plant height variability by +/- Hr * height}

\item{reps}{Number of repetitions}

\item{l}{Variation around input leaf dimensions}

\item{Ms}{Standard deviation of LFMC}

\item{Pm}{Multiplier of mean LFMC}

\item{Mr}{Truncates LFMC variability by +/- Pm * LFMC}
}
\value{
dataframe
}
\description{
Calculates a non-deterministic wind profile on parallel cores
Useful when running a large number of reps
}
