% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Shelter.R
\name{WRF}
\alias{WRF}
\title{Calculates the wind reduction factor from a single parameter file,
defaults to screen height of 1.2m}
\usage{
WRF(base.params, test = 1.2)
}
\arguments{
\item{base.params}{Parameter input table}

\item{test}{The lower height in m above ground}
}
\value{
numeric
}
\description{
Calculates the wind reduction factor from a single parameter file,
defaults to screen height of 1.2m
}
