% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Growth.R
\name{applyGrowth}
\alias{applyGrowth}
\title{Updates a parameter file with the growth parameters modelled for a species}
\usage{
applyGrowth(Param, sp, current)
}
\arguments{
\item{Param}{Parameter file}

\item{sp}{Name of the species being grown}

\item{current}{dataframe produced by growPlants}
}
\value{
dataframe
}
\description{
Note: w is currently not changed due to an error:
"Error in .match_param(param, section, no.match.error = TRUE, single = TRUE) :
w matches more than one parameter"
}
