% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impact.R
\name{IMPACT}
\alias{IMPACT}
\title{The Impact of Items}
\usage{
IMPACT(x, y)
}
\arguments{
\item{x}{is a data set}

\item{y}{is a null value}
}
\value{
IMPACT a summary table with the impact of items
}
\description{
This function returns an estimation based on the patterns of items.


Introduce a set of categorical data set classifed as numerical data
}
\details{
This package runs a multivariate analysis of the impact of items to identify a bias in the questionnaire validation of likert-type scale variables. The items requieres considering a null value (category doesn't have tendency). Offering frequency, importance and impact of the items.
It estimates the impact of items

This funtion takes a set of values produced by the IMPACT
functions returns estimations for each item provided in the input x matrix.
}
\examples{
\dontrun{
library(IMPACT)
##Reads a likert-type scale dataset
x<-matrix(c(2, 5, 5, 4, 4, 5, 4, 4, 5, 1, 3, 4, 4, 1, 5, 2, 2, 4, 3, 5,
5, 1, 1, 4, 5, 2, 2, 4, 4, 5, 2, 4, 2, 5, 3, 4, 3, 3, 5, 3,
3, 5, 5, 5, 5, 5, 5, 5, 5, 3, 4, 3, 5, 3, 4, 5, 4, 4, 4, 5,
4, 1, 2, 2, 3, 1, 1, 3, 2, 5, 3, 2, 1, 5, 2, 2, 4, 1, 5, 1,
2, 4, 4, 4, 3, 5, 5, 4, 2, 2, 4, 3, 5, 2, 4, 5, 4, 4, 1, 5,
4, 1, 2, 3, 3, 1, 2, 5, 4, 5, 4, 3, 1, 4, 1, 3, 4, 2, 4, 2,
4, 1, 2, 3, 4, 1, 1, 4, 4, 5, 3, 3, 1, 1, 1, 4, 4, 2, 4, 1,
5, 1, 3, 3, 4, 5, 3, 5, 4, 5, 4, 4, 2, 5, 2, 4, 4, 4, 4, 4,
4, 4, 2, 3, 1, 3, 2, 3, 3, 1, 1, 2, 4, 1, 5, 2, 2, 3, 3, 4),20,10)
##Put names of variables
colnames(x)<-c(paste("A","-",1:10))
##Declare a null value
y<-3
IMPACT(x,y)
}

}
\author{
Nery Sofia Huerta-Pacheco
}
\references{
Juniper, E. F., Guyatt, G. H., Streiner, D. L., & King, D. R. (1997). Clinical impact
versus factor analysis for quality of life questionnaire construction. Journal of
ClinicalEpidemiology, 233-238.

Allen, F., & Locker, D. (2002). A Modified Short Version of the Oral Health Impact
Profile for Assessing Health-Related Quality of Life in Edentulous Adults. The
International Journal of Prosthodontics, 15(5), 446-450.

Vicente Galindo, E. D. (2011). Analisis del Impacto frente a Teoria de Respuesta
al Item (Trabajo Fin de Master). Master Universitario en Analisis Avanzado de Datos
Multivariantes, Statistics Department, University of Salamanca, Spain.
}

