% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ensembles.R
\name{plantVar}
\alias{plantVar}
\title{Randomly modifies plant traits within defined ranges for non-deterministic predictions}
\usage{
plantVar(base.params, Strata, Species, l = 0.1, Ms = 0.01, Pm = 1,
  Mr = 1.001, Hs = 0.2, Hr = 1.41)
}
\arguments{
\item{base.params}{Parameter input table}

\item{Strata}{Strata descriptor table output by the function 'strata'}

\item{Species}{Species descriptor table output by the function 'species'}

\item{l}{Variation around input leaf dimensions}

\item{Ms}{Standard deviation of LFMC}

\item{Pm}{Multiplier of mean LFMC}

\item{Mr}{Truncates LFMC variability by +/- Pm * LFMC}

\item{Hs}{Standard deviation of plant height variations}

\item{Hr}{Truncates plant height variability by +/- Hr * height}
}
\value{
dataframe
}
\description{
Randomly modifies plant traits within defined ranges for non-deterministic predictions
}
