% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_impact_data_step_change.R
\name{plot_impact_data_step_change}
\alias{plot_impact_data_step_change}
\title{Plot the impact data step change}
\usage{
plot_impact_data_step_change(dat, method = "diff", type = "raw",
  val_name = "val")
}
\arguments{
\item{dat}{impact data}

\item{method}{diff or ratio}

\item{type}{raw or corrected}

\item{val_name}{the name of the value column}
}
\value{
ggplot2 object
}
\description{
Plot the impact data step change
}
\examples{
dat <- data.frame(trt = rep(c("test","ctrl"), each = 6),
                  pre_post = rep(c("pre","post"), each = 3, times = 2),
                  id = as.character(c(1,2,3,1,2,3,4,5,6,4,5,6)),
                  time = as.character(c(1,1,1,2,2,2,1,1,1,2,2,2)),
                  val = c(10,11,12,13,14,15,10,11,10,12,10,10.5))
plot_impact_data_step_change(dat)
plot_impact_data_step_change(dat, type = "raw")
}
