% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_impactResult.R
\name{print.impactResult}
\alias{print.impactResult}
\title{Print Impact Result}
\usage{
\method{print}{impactResult}(x, ...)
}
\arguments{
\item{x}{impactResult object}

\item{...}{future parameters}
}
\description{
Print Impact Result
}
\examples{
a <- 4
}
