% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Summaries.R
\name{species}
\alias{species}
\title{Species descriptors from a param file}
\usage{
species(base.params)
}
\arguments{
\item{base.params}{Input parameter file}
}
\value{
dataframe
}
\description{
Finds dimensions and moisture of each species
}
