% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Summaries.R
\name{strata}
\alias{strata}
\title{Stratum descriptors from a param file}
\usage{
strata(base.params)
}
\arguments{
\item{base.params}{Input parameter file}
}
\value{
dataframe
}
\description{
For each stratum, finds mean crown width, plant separation, and number of species
}
