% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Summaries.R
\name{summary}
\alias{summary}
\title{Summary table of fire behaviour}
\usage{
summary(Stratum, Surf)
}
\arguments{
\item{Stratum}{The dataframe produced by the extension stratum}

\item{Surf}{The dataframe produced by the extension surf}
}
\value{
dataframe
}
\description{
Summarises FRaME generated fire behaviour by RepId
}
